[REFAC] Update site to only have 1 courses/show
There is no need for the courses show for regular users to be
different from the admin view.  One or two conditionals can cover the
needed additional links.

This commit also adds a view test for the edit course link, and fixes
other tests that were leaving users behind in the database.