[CHG] Updates js event for completed lesson
When an ASL file is finished, we have control over the JavaScript
methods that are called upon completion. This updates the method name
to be inline with what was agreed upon with Jamie from the ASL file.

Also, this removes the temporary button needed in development to fire
the same event.

A bug was fixed where the Modal was not being closed before going to
the next lesson, so hitting the back button still showed the modal,
which causes issues.