[CHG] Allow rendering of search box as partial
This breaks out the search box into its own partial. It also sets up
the 8/4 column divs. I used a &nbsp; to make the 8 column divs actually
appear. I’m not sure if there is a better way to do this with neat. In
bootstrap, I’ve used an offset.