[CHG] Adds hide/show password functionality
Adds a plugin to manage hiding and showing of passwords. Converts the
text fields we were using for passwords to password fields, and then
adds a checkbox for ‘show password’, which defaults to checked. On
checkbox change, toggles the password field to either show or not based
on the value of the checkbox.