[CHG] Creates organization course
This will allow us to connect courses to organizations. Currently, a
course only has one org, but that won’t always be the case, and this is
setup to handle that. I envision a boolean that allows us to mark an
organization_course as a “copy” and then we could use that to allow
orgs to share courses amongst each other.