[CHG] Adds chicago redirect
This is needed as the CNAME only points to the right server. With this,
they will get redirected to the correct subdomain, which is what we
need. In the future, we can use this with a bit of metaprogramming to
allow people to setup a “main” subdomain and “secondary” subdomains
that redirect to their main subdomain.