[CHG] Adds conditional abilities to Save and Add/Edit course button
This does a few things. First, it adds a conditional to change the
language on the button. Then, it add a case statement to match against
that submit language on the button to redirect to the right place.
Finally, it updates the test so that it can find the button with the
new name.